Split resolver type behavior metadata tests
