+++ src/typechecker/tests/resolver_type_behavior_metadata/behavior_symbols.rs
@@ -1,0 +1,143 @@
+use super::*;
+
+#[test]
+fn check_program_with_symbols_validates_resolver_behavior_visibility() {
+    let program = parse_program(
+        r#"
+Json: behavior {
+    encode: (Self) StaticString
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_public_for_test(Namespace::Behavior, "Json", true);
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver behavior visibility mismatch should fail");
+
+    assert!(
+        err.iter().any(|d| d
+            .message
+            .contains("resolver behavior symbol 'Json' has visibility public, expected private")),
+        "expected resolver behavior visibility diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_behavior_type_parameter_bounds() {
+    let program = parse_program(
+        r#"
+Json<T>: behavior {
+    encode: (Self) T
+}
+Serializable<T: Json<T>>: behavior {
+    serialize: (Self) T
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_type_parameter_bounds_for_test(
+        Namespace::Behavior,
+        "Serializable",
+        Some(vec![("T".to_string(), "Json<i32>".to_string())]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver behavior generic bound mismatch should fail");
+
+    assert!(
+            err.iter().any(|d| d.message.contains(
+                "resolver behavior symbol 'Serializable' has type parameter bounds '(T: Json<i32>)', expected '(T: Json<T>)'"
+            )),
+            "expected resolver behavior generic bound diagnostic, got {err:?}"
+        );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_behavior_absent_type_metadata() {
+    let program = parse_program(
+        r#"
+Json: behavior {
+    encode: (Self) StaticString
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_field_count_for_test(Namespace::Behavior, "Json", Some(1));
+    symbols.set_field_type_names_for_test(
+        Namespace::Behavior,
+        "Json",
+        Some(vec![("value".to_string(), "i32".to_string())]),
+    );
+    symbols.set_field_types_for_test(
+        Namespace::Behavior,
+        "Json",
+        Some(vec![("value".to_string(), AstType::I32)]),
+    );
+    symbols.set_variant_names_for_test(Namespace::Behavior, "Json", Some(vec!["Some".to_string()]));
+    symbols.set_variant_payload_type_name_for_test(
+        Namespace::Behavior,
+        "Json",
+        Some("i32".to_string()),
+    );
+    symbols.set_variant_payload_type_for_test(Namespace::Behavior, "Json", Some(AstType::I32));
+    symbols.set_behavior_impl_names_for_test(
+        Namespace::Behavior,
+        "Json",
+        Some(vec!["Debug".to_string()]),
+    );
+    symbols.set_behavior_impl_refs_for_test(
+        Namespace::Behavior,
+        "Json",
+        Some(vec![BehaviorRefMetadata {
+            name: "Debug".to_string(),
+            type_args: Vec::new(),
+        }]),
+    );
+    symbols.set_behavior_required_names_for_test(
+        Namespace::Behavior,
+        "Json",
+        Some(vec!["Debug".to_string()]),
+    );
+    symbols.set_behavior_required_refs_for_test(
+        Namespace::Behavior,
+        "Json",
+        Some(vec![BehaviorRefMetadata {
+            name: "Debug".to_string(),
+            type_args: Vec::new(),
+        }]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver behavior type metadata should fail");
+
+    for expected in [
+        "resolver behavior symbol 'Json' has field count metadata, expected none",
+        "resolver behavior symbol 'Json' has field types metadata, expected none",
+        "resolver behavior symbol 'Json' has typed field types metadata, expected none",
+        "resolver behavior symbol 'Json' has variant names metadata, expected none",
+        "resolver behavior symbol 'Json' has variant payload type metadata, expected none",
+        "resolver behavior symbol 'Json' has typed variant payload type metadata, expected none",
+        "resolver behavior symbol 'Json' has behavior impls metadata, expected none",
+        "resolver behavior symbol 'Json' has typed behavior impls metadata, expected none",
+        "resolver behavior symbol 'Json' has behavior requires metadata, expected none",
+        "resolver behavior symbol 'Json' has typed behavior requires metadata, expected none",
+    ] {
+        assert!(
+            err.iter().any(|d| d.message.contains(expected)),
+            "expected resolver behavior type metadata diagnostic '{expected}', got {err:?}"
+        );
+    }
+}
